# Optimize file reader memory footprint