--- a/buffer_test.go
+++ b/buffer_test.go
@@ -371,7 +371,9 @@ func TestBufferGenerateBloomFilters(t *testing.T) {
 		for i := range rows {
 			buffer.Write(&rows[i])
 		}
-		_, err := parquet.CopyRows(writer, buffer.Rows())
+		bufferRows := buffer.Rows()
+		_, err := parquet.CopyRows(writer, bufferRows)
+		bufferRows.Close()
 		if err != nil {
 			t.Error(err)
 			return false