--- a/merge.go
+++ b/merge.go
@@ -19,10 +19,10 @@ func (m *mergedRowGroup) SortingColumns() []SortingColumn {
 func (m *mergedRowGroup) Rows() Rows {
 	// The row group needs to respect a sorting order; the merged row reader
 	// uses a heap to merge rows from the row groups.
-	return &mergedRowGroupRowReader{rowGroup: m, schema: m.schema}
+	return &mergedRows{rowGroup: m, schema: m.schema}
 }
 
-type mergedRowGroupRowReader struct {
+type mergedRows struct {
 	rowGroup *mergedRowGroup
 	schema   *Schema
 	sorting  []columnSortFunc
@@ -34,7 +34,7 @@ type mergedRowGroupRowReader struct {
 	err      error
 }
 
-func (r *mergedRowGroupRowReader) init(m *mergedRowGroup) {
+func (r *mergedRows) init(m *mergedRowGroup) {
 	if r.schema != nil {
 		numColumns := numLeafColumnsOf(r.schema)
 		cursors := make([]bufferedRowGroupCursor, len(m.rowGroups))
@@ -53,14 +53,14 @@ func (r *mergedRowGroupRowReader) init(m *mergedRowGroup) {
 			// TODO: this is a bit of a weak model, it only works with types
 			// declared in this package; we may want to define an API to allow
 			// applications to participate in it.
-			if rd, ok := cursors[i].reader.(*rowGroupRowReader); ok {
+			if rd, ok := cursors[i].reader.(*rowGroupRows); ok {
 				rd.init(rd.rowGroup)
 				rd.rowGroup = nil
 				// TODO: this optimization is disabled for now, there are
 				// remaining blockers:
 				//
 				// * The optimized merge of pages for non-overlapping ranges is
-				//   not yet implemented in the mergedRowGroupRowReader type.
+				//   not yet implemented in the mergedRows type.
 				//
 				// * Using pages min/max to determine overlapping ranges does
 				//   not work for repeated columns; sorting by repeated columns
@@ -86,7 +86,14 @@ func (r *mergedRowGroupRowReader) init(m *mergedRowGroup) {
 	}
 }
 
-func (r *mergedRowGroupRowReader) SeekToRow(rowIndex int64) error {
+func (r *mergedRows) Close() error {
+	for i := range r.cursors {
+		r.cursors[i].close()
+	}
+	return nil
+}
+
+func (r *mergedRows) SeekToRow(rowIndex int64) error {
 	if rowIndex >= r.index {
 		r.seek = rowIndex
 		return nil
@@ -94,7 +101,7 @@ func (r *mergedRowGroupRowReader) SeekToRow(rowIndex int64) error {
 	return fmt.Errorf("SeekToRow: merged row reader cannot seek backward from row %d to %d", r.index, rowIndex)
 }
 
-func (r *mergedRowGroupRowReader) ReadRow(row Row) (Row, error) {
+func (r *mergedRows) ReadRow(row Row) (Row, error) {
 	if r.rowGroup != nil {
 		r.init(r.rowGroup)
 		r.rowGroup = nil
@@ -131,7 +138,7 @@ func (r *mergedRowGroupRowReader) ReadRow(row Row) (Row, error) {
 	}
 }
 
-// func (r *mergedRowGroupRowReader) WriteRowsTo(w RowWriter) (int64, error) {
+// func (r *mergedRows) WriteRowsTo(w RowWriter) (int64, error) {
 // 	if r.rowGroup != nil {
 // 		defer func() { r.rowGroup = nil }()
 // 		switch dst := w.(type) {
@@ -145,22 +152,22 @@ func (r *mergedRowGroupRowReader) ReadRow(row Row) (Row, error) {
 // 	return CopyRows(w, struct{ RowReaderWithSchema }{r})
 // }
 
-func (r *mergedRowGroupRowReader) writeRowsTo(w pageAndValueWriter) (numRows int64, err error) {
+func (r *mergedRows) writeRowsTo(w pageAndValueWriter) (numRows int64, err error) {
 	// TODO: the intent of this method is to optimize the merge of rows by
 	// copying entire pages instead of individual rows when we detect ranges
 	// that don't overlap.
 	return
 }
 
-func (r *mergedRowGroupRowReader) Schema() *Schema {
+func (r *mergedRows) Schema() *Schema {
 	return r.schema
 }
 
-func (r *mergedRowGroupRowReader) Len() int {
+func (r *mergedRows) Len() int {
 	return len(r.cursors)
 }
 
-func (r *mergedRowGroupRowReader) Less(i, j int) bool {
+func (r *mergedRows) Less(i, j int) bool {
 	cursor1 := r.cursors[i]
 	cursor2 := r.cursors[j]
 
@@ -179,15 +186,15 @@ func (r *mergedRowGroupRowReader) Less(i, j int) bool {
 	return false
 }
 
-func (r *mergedRowGroupRowReader) Swap(i, j int) {
+func (r *mergedRows) Swap(i, j int) {
 	r.cursors[i], r.cursors[j] = r.cursors[j], r.cursors[i]
 }
 
-func (r *mergedRowGroupRowReader) Push(interface{}) {
+func (r *mergedRows) Push(interface{}) {
 	panic("BUG: unreachable")
 }
 
-func (r *mergedRowGroupRowReader) Pop() interface{} {
+func (r *mergedRows) Pop() interface{} {
 	n := len(r.cursors) - 1
 	c := r.cursors[n]
 	r.cursors = r.cursors[:n]
@@ -195,6 +202,7 @@ func (r *mergedRowGroupRowReader) Pop() interface{} {
 }
 
 type rowGroupCursor interface {
+	close() error
 	readRow(Row) (Row, error)
 	readNext() error
 	nextRowValuesOf([]Value, int16) []Value
@@ -209,6 +217,10 @@ type bufferedRowGroupCursor struct {
 	reader  Rows
 	rowbuf  Row
 	columns [][]Value
+}
+
+func (cur *bufferedRowGroupCursor) close() error {
+	return cur.reader.Close()
 }
 
 func (cur *bufferedRowGroupCursor) readRow(row Row) (Row, error) {
@@ -235,7 +247,7 @@ func (cur *bufferedRowGroupCursor) nextRowValuesOf(values []Value, columnIndex i
 }
 
 /*
-type optimizedRowGroupCursor struct{ *rowGroupRowReader }
+type optimizedRowGroupCursor struct{ *rowGroupRows }
 
 func (cur optimizedRowGroupCursor) readRow(row Row) (Row, error) { return cur.ReadRow(row) }
 
@@ -264,6 +276,6 @@ func (cur optimizedRowGroupCursor) nextRowValuesOf(values []Value, columnIndex i
 */
 
 var (
-	_ RowReaderWithSchema = (*mergedRowGroupRowReader)(nil)
-	//_ RowWriterTo         = (*mergedRowGroupRowReader)(nil)
+	_ RowReaderWithSchema = (*mergedRows)(nil)
+	//_ RowWriterTo         = (*mergedRows)(nil)
 )