This is still a work in progress but generally ready to get feedback I think.

Benchmarks were showing that reading parquet files was allocating large amounts of memory, so I spent a bit of time optimizing it.

I took the following approaches:
* Make the size of internal buffers configurable when reading files, which helps applications reduce the amount of memory they allocate if needed
* Fix a bug that was causing allocations of read buffers to be multiplied by the size of the type (e.g. 8x for a buffer of int64)
* pool the memory buffer used to read compressed pages into memory

These changes have been effective at reducing the amount of memory allocated by the program:
```
name                 old time/op    new time/op    delta
ReaderReadRow/INT64    16.7ns ± 0%    16.1ns ± 1%   -3.81%  (p=0.000 n=9+9)

name                 old speed      new speed      delta
ReaderReadRow/INT64   538MB/s ± 0%   560MB/s ± 1%   +3.96%  (p=0.000 n=9+9)

name                 old row/s      new row/s      delta
ReaderReadRow/INT64     59.8M ± 0%     62.2M ± 1%   +3.96%  (p=0.000 n=9+9)

name                 old alloc/op   new alloc/op   delta
ReaderReadRow/INT64     12.0B ± 0%      1.0B ± 0%  -91.67%  (p=0.000 n=10+10)

name                 old allocs/op  new allocs/op  delta
ReaderReadRow/INT64      0.00           0.00          ~     (all equal)
```

In order to enable pooling of memory buffers, I would like to modify some exported APIs to give the program a mechanism for controlling the lifecycle of some objects (e.g. page readers) via a call to a `Close` method when the objects are not needed anymore. This is an area where I am still experimenting, I may extend this to more APIs to support better internal memory management.