@achille-roussel 2022-05-17
Closing in favor of https://github.com/segmentio/parquet-go/pull/182